pyproject.toml
[project]
name = "edgewatch-telemetry"
version = "0.4.5"
description = "Local-first edge telemetry platform: device ingestion, buffering, status, alerts, and a small dashboard."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Ryne Schroder" }]

dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.30",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "apscheduler>=3.10",
  "python-dotenv>=1.0",
  "requests>=2.31",
  "alembic>=1.13",
  "pyyaml>=6.0",
]

[dependency-groups]
# Dev tooling for harness tasks and CI.
#
# Note: We prefer Pyright for typechecking (aligns with harness scaffold),
# but keep mypy available as an optional secondary checker.
dev = [
  "pre-commit>=3.0",
  "ruff>=0.5",
  "pyright>=1.1.0",
  "pytest>=8.0",
  "mypy>=1.10",
  "types-PyYAML>=6.0",
]

[tool.ruff]
line-length = 110
target-version = "py311"
extend-exclude = [
  ".venv",
  ".pytest_cache",
  ".ruff_cache",
  "infra/gcp/.terraform",
  "web/dist",
  "web/node_modules",
]

[tool.ruff.format]
quote-style = "double"

[tool.pyright]
# Keep Pyright "strict"-ish, but avoid busywork in SQLAlchemy-heavy areas.
# Tighten over time as the project grows.
typeCheckingMode = "basic"
pythonVersion = "3.11"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.pytest.ini_options]
# Ensure local imports work in a "flat" repo layout (api/, agent/, etc.).
addopts = ["--import-mode=prepend"]
testpaths = ["tests"]
pythonpath = ["."]

[tool.uv]
# Guardrail: require a *minimum* uv version that supports the lockfile format.
#
# Do NOT pin an exact version here unless you control all environments, otherwise
# contributors will hit confusing failures if their `uv` is slightly older/newer.
required-version = ">=0.9.0"
